requirements.txt
Django>=4.2
gunicorn>=20.1.0
psycopg2-binary>=2.9
python-magic>=0.4.27
python-decouple>=3.8
requests>=2.28
celery>=5.2
redis>=4.5
django-redis==5.4.0
openai>=0.27.0
python-dotenv==1.0.1
pefile>=2023.2.7